[util] Extract shortest path calculation into a new module
